+++ systems/identity/identity-application/src/main/kotlin/hs/kr/entrydsm/identity/application/port/in/command/SignupCommand.kt
@@ -1,6 +1,6 @@
 package hs.kr.entrydsm.identity.application.port.`in`.command
 
-import hs.kr.entrydsm.identity.domain.SignupType
+import hs.kr.entrydsm.identity.domain.enum.SignupType
 import java.time.LocalDate
 
 data class SignupCommand(
