+++ systems/identity/identity-application/src/main/kotlin/hs/kr/entrydsm/identity/application/port/in/result/AccountResult.kt
@@ -1,6 +1,6 @@
 package hs.kr.entrydsm.identity.application.port.`in`.result
 
-import hs.kr.entrydsm.identity.domain.AccountStatus
+import hs.kr.entrydsm.identity.domain.enum.AccountStatus
 import java.time.Instant
 
 data class AccountResult(
