+++ systems/identity/identity-application/src/main/kotlin/hs/kr/entrydsm/identity/application/port/in/result/UserSummaryResult.kt
@@ -1,6 +1,6 @@
 package hs.kr.entrydsm.identity.application.port.`in`.result
 
-import hs.kr.entrydsm.identity.domain.AccountStatus
+import hs.kr.entrydsm.identity.domain.enum.AccountStatus
 
 data class UserSummaryResult(
     val userId: Long,
